// frontend/src/components/types.ts
export interface Product {
  id: string;
  name: string;
  purchasePrice: number;
  taxRate: number;
  breakageRate: number;
  agingRate: number;
  mixAvailable: number;
}